Move DTOs and NetworkClient to appropriate packages
